Create location.dal.getAttachments
