+++ server/api/locations/location/dal/index.js
@@ -3,6 +3,7 @@ exports.changeName = require('./changeName');
 exports.changeStatus = require('./changeStatus');
 exports.changeType = require('./changeType');
 exports.changeThumbnail = require('./changeThumbnail');
+exports.getAttachments = require('./getAttachments');
 exports.getOneComplete = require('./getOneComplete');
 exports.getRaw = require('./getRaw');
 exports.removeOne = require('./removeOne');
